fix_event_activation_by_role
Add functionality to modify active status of event role and display/handle roles with different active statuses in RolesAndProfilesScreen

- Add import for dart:convert, FlutterFlowUtil, and TempDataFunctions
- Modify the onChanged callback for SwitchListTile in RolesAndProfilesScreen to async and call modifyActiveOfEventRole function
- Add modifyActiveOfEventRole function in api\_calls\_list.dart
- Add if conditions to display and handle icons for activating and deactivating roles in RolesAndProfilesScreen
- Comment out the previous ListView.builder code for displaying grouped events

Add modifyActiveOfEventRole function in api\_calls\_list.dart
--------------------------------------------------------------

Add a new function modifyActiveOfEventRole in api\_calls\_list.dart to modify the active status of an event role. This function takes in three parameters: eventId, role\_event\_value, and role\_selected. It creates a body object with these parameters and sends a PUT request to the specified API endpoint with the required headers and body. If the request is successful, it returns the content of the response. If there is an exception, it throws a FormatException with the error message.

Modify the onChanged callback for SwitchListTile in RolesAndProfilesScreen to async and call modifyActiveOfEventRole function
----------------------------------------------------------------------------------------------------------------------------------

Modify the onChanged callback for SwitchListTile in RolesAndProfilesScreen to async and call modifyActiveOfEventRole function to modify the active status of an event role.

Add if conditions to display and handle icons for activating and deactivating roles in RolesAndProfilesScreen
----------------------------------------------------------------------------------------------------------

Add if conditions to display and handle icons for activating and deactivating roles in RolesAndProfilesScreen based on the active status of the role.

Comment out the previous ListView.builder code for displaying grouped events
-----------------------------------------------------------------------------

Comment out the previous ListView.builder code for displaying grouped events in RolesAndProfilesScreen as it is not used in the current implementation.